Fixed 'alert.priority' not lookup base on overridden search result
As stated in 'Add action':
"Default impact for incidents of this alert.
Note: The impact can be overriden by a field from search results named 'impact'. Later, the alert manager calculates a priority based on the impact and urgency."

That is not the case. Orginally, the result.impact and result.urgency is not used to lookup alert.priority, thus, causing other errors. For example, $alert.priority$ in email is using the the payload.configuration.impact & payload.configuration.urgency for lookup.