Align time to show trends more clearly
Default timechart behaviour aligns time to midnight. This causes trend lines to only represent a part of a day, and consequently almost always appear as a downward trend regardless of what is actually occurring. Using aligntime and defining span in minutes instead of days allows it to display as a representation of the last rolling 24h compared to the 24h span preceding that.